requirements.txt
python-dotenv
requests
wit
